+++ age/src/plugin.rs
@@ -199,22 +199,19 @@ impl fmt::Display for Identity {
 
 impl Identity {
     /// Returns the identity corresponding to the given plugin name in its default mode.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `plugin_name` contains invalid characters.
-    pub fn default_for_plugin(plugin_name: &str) -> Self {
+    pub fn default_for_plugin(plugin_name: &str) -> Result<Self, ResolveError> {
         if valid_plugin_name(plugin_name) {
-            bech32_encode(
+            Ok(bech32_encode(
                 Hrp::parse_unchecked(&format!("{}{}-", PLUGIN_IDENTITY_PREFIX, plugin_name)),
                 &[],
             )
             .to_uppercase()
             .parse()
-            .unwrap()
+            .unwrap())
         } else {
-            // TODO: Change the API to be fallible.
-            panic!("invalid plugin name")
+            Err(ResolveError::InvalidPluginName {
+                name: plugin_name.into(),
+            })
         }
     }
 
@@ -232,9 +229,11 @@ pub(crate) struct Plugin {
 
 impl Plugin {
     /// Finds the age plugin with the given name in `$PATH`.
-    ///
-    /// On error, returns the binary name that could not be located.
-    pub(crate) fn new(name: &str) -> Result<Self, String> {
+    pub(crate) fn new(name: &str) -> Result<Self, ResolveError> {
+        if !valid_plugin_name(name) {
+            return Err(ResolveError::InvalidPluginName { name: name.into() });
+        }
+
         let binary_name = binary_name(name);
         match which::which(&binary_name).or_else(|e| {
             // If we are running in WSL, try appending `.exe`; plugins installed in
@@ -247,7 +246,7 @@ impl Plugin {
             }
         }) {
             Ok(path) => Ok(Plugin { binary_name, path }),
-            Err(_) => Err(binary_name),
+            Err(_) => Err(ResolveError::MissingPlugin { binary_name }),
         }
     }
 
@@ -393,29 +392,21 @@ impl<C: Callbacks> RecipientPluginV1<C> {
         recipients: &[Recipient],
         identities: &[Identity],
         callbacks: C,
-    ) -> Result<Self, EncryptError> {
-        if valid_plugin_name(plugin_name) {
-            Plugin::new(plugin_name)
-                .map_err(|binary_name| EncryptError::MissingPlugin { binary_name })
-                .map(|plugin| RecipientPluginV1 {
-                    plugin,
-                    recipients: recipients
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect(),
-                    identities: identities
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect(),
-                    callbacks,
-                })
-        } else {
-            Err(EncryptError::MissingPlugin {
-                binary_name: plugin_name.to_string(),
-            })
-        }
+    ) -> Result<Self, ResolveError> {
+        Plugin::new(plugin_name).map(|plugin| RecipientPluginV1 {
+            plugin,
+            recipients: recipients
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect(),
+            identities: identities
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect(),
+            callbacks,
+        })
     }
 }
 
@@ -603,23 +594,15 @@ impl<C: Callbacks> IdentityPluginV1<C> {
         plugin_name: &str,
         identities: &[Identity],
         callbacks: C,
-    ) -> Result<Self, DecryptError> {
-        if valid_plugin_name(plugin_name) {
-            Plugin::new(plugin_name)
-                .map_err(|binary_name| DecryptError::MissingPlugin { binary_name })
-                .map(|plugin| {
-                    let identities = identities
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect();
-                    Self::from_parts(plugin, identities, callbacks)
-                })
-        } else {
-            Err(DecryptError::MissingPlugin {
-                binary_name: plugin_name.to_string(),
-            })
-        }
+    ) -> Result<Self, ResolveError> {
+        Plugin::new(plugin_name).map(|plugin| {
+            let identities = identities
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect();
+            Self::from_parts(plugin, identities, callbacks)
+        })
     }
 
     pub(crate) fn from_parts(plugin: Plugin, identities: Vec<Identity>, callbacks: C) -> Self {
@@ -742,16 +725,48 @@ impl<C: Callbacks> crate::Identity for IdentityPluginV1<C> {
     }
 }
 
+/// Errors returned when resolving a plugin.
+#[derive(Clone, Debug)]
+#[non_exhaustive]
+pub enum ResolveError {
+    /// A provided plugin name was invalid.
+    InvalidPluginName {
+        /// The provided name.
+        name: String,
+    },
+    /// A required plugin could not be found.
+    MissingPlugin {
+        /// The plugin's binary name.
+        binary_name: String,
+    },
+}
+
+impl fmt::Display for ResolveError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ResolveError::InvalidPluginName { name } => {
+                wfl!(f, "err-invalid-plugin-name", name = name.as_str())
+            }
+            ResolveError::MissingPlugin { binary_name } => {
+                wlnfl!(f, "err-missing-plugin", plugin_name = binary_name.as_str())?;
+                wfl!(f, "rec-missing-plugin")
+            }
+        }
+    }
+}
+
+impl std::error::Error for ResolveError {}
+
 #[cfg(test)]
 mod tests {
     use age_core::primitives::bech32_encode;
     use bech32::Hrp;
 
-    use crate::{DecryptError, EncryptError, NoCallbacks};
+    use crate::NoCallbacks;
 
     use super::{
-        Identity, IdentityPluginV1, Recipient, RecipientPluginV1, PLUGIN_IDENTITY_PREFIX,
-        PLUGIN_RECIPIENT_PREFIX,
+        Identity, IdentityPluginV1, Recipient, RecipientPluginV1, ResolveError,
+        PLUGIN_IDENTITY_PREFIX, PLUGIN_RECIPIENT_PREFIX,
     };
 
     const INVALID_PLUGIN_NAME: &str = "foobar/../../../../../../../usr/bin/echo";
@@ -759,7 +774,7 @@ mod tests {
     #[test]
     fn default_for_plugin() {
         assert_eq!(
-            Identity::default_for_plugin("foobar").to_string(),
+            Identity::default_for_plugin("foobar").unwrap().to_string(),
             "AGE-PLUGIN-FOOBAR-1QVHULF",
         );
     }
@@ -806,22 +821,26 @@ mod tests {
     }
 
     #[test]
-    #[should_panic]
     fn identity_default_for_plugin_rejects_empty_name() {
-        Identity::default_for_plugin("");
+        assert!(matches!(
+            Identity::default_for_plugin(""),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
+        ));
     }
 
     #[test]
-    #[should_panic]
     fn identity_default_for_plugin_rejects_invalid_chars() {
-        Identity::default_for_plugin(INVALID_PLUGIN_NAME);
+        assert!(matches!(
+            Identity::default_for_plugin(INVALID_PLUGIN_NAME),
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
+        ));
     }
 
     #[test]
     fn recipient_plugin_v1_rejects_empty_name() {
         assert!(matches!(
             RecipientPluginV1::new("", &[], &[], NoCallbacks),
-            Err(EncryptError::MissingPlugin { binary_name }) if binary_name.is_empty(),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
         ));
     }
 
@@ -829,7 +848,7 @@ mod tests {
     fn recipient_plugin_v1_rejects_invalid_chars() {
         assert!(matches!(
             RecipientPluginV1::new(INVALID_PLUGIN_NAME, &[], &[], NoCallbacks),
-            Err(EncryptError::MissingPlugin { binary_name }) if binary_name == INVALID_PLUGIN_NAME,
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
         ));
     }
 
@@ -837,7 +856,7 @@ mod tests {
     fn identity_plugin_v1_rejects_empty_name() {
         assert!(matches!(
             IdentityPluginV1::new("", &[], NoCallbacks),
-            Err(DecryptError::MissingPlugin { binary_name }) if binary_name.is_empty(),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
         ));
     }
 
@@ -845,7 +864,7 @@ mod tests {
     fn identity_plugin_v1_rejects_invalid_chars() {
         assert!(matches!(
             IdentityPluginV1::new(INVALID_PLUGIN_NAME, &[], NoCallbacks),
-            Err(DecryptError::MissingPlugin { binary_name }) if binary_name == INVALID_PLUGIN_NAME,
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
         ));
     }
 }